% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_formulas.R
\name{BCRA3_2f}
\alias{BCRA3_2f}
\title{@title Model 4.1:  MDES Calculator for 3-Level Fixed Effects Blocked Cluster Random Assignment Design (BCRA3_2f)—Treatment at Level 2
 @description MDES for 3-Level Fixed Effects Blocked Cluster Random Assignment Design}
\usage{
BCRA3_2f(R2_1, R2_2, P, g_2, n, J, K, rho_2, alpha = 0.05, power = 0.8,
  tails = 2)
}
\arguments{
\item{R2_1}{Proportion of variance in level 1 outcome explained by level 1 covariates}

\item{R2_2}{Proportion of variance in level 2 outcome explained by level 2 covariates}

\item{P}{Proportion of the level 2 units randomized to the treatment}

\item{g_2}{Number of level 2 Covariates used}

\item{n}{Mean # of level 1 units per level 2 unit}

\item{J}{Mean # of level 2 units per level 3 unit}

\item{K}{Number of level 3 units}

\item{rho_2}{proportion of variance in outcome that is between level 2 units}

\item{alpha}{Probability of a type I error}

\item{power}{Statistical power (1 - probability of type II error)}

\item{tails}{One or two tailed test}
}
\value{
Minimum Detectable Effect Size
}
\description{
@title Model 4.1:  MDES Calculator for 3-Level Fixed Effects Blocked Cluster Random Assignment Design (BCRA3_2f)—Treatment at Level 2
 @description MDES for 3-Level Fixed Effects Blocked Cluster Random Assignment Design
}

