% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_formulas.R
\name{BIRA2_1c}
\alias{BIRA2_1c}
\title{Model 2.1:  MDES Calculator for 2-Level Constant Effects Blocked Individual Random Assignment Designs (BIRA2_1c)—School Intercepts Only}
\usage{
BIRA2_1c(R2_1, P, g, n, J, alpha = 0.05, power = 0.8, tails = 2)
}
\arguments{
\item{R2_1}{Proportion of variance in level 1 outcome explained by block and level 1 covariates}

\item{P}{Proportion of the level 1 units randomized to the treatment}

\item{g}{Number of level 1 Covariates used}

\item{n}{Average Block Size (mean # of level 1 units per level 2 unit)}

\item{J}{Number of level 2 units}

\item{alpha}{Probability of a type I error}

\item{power}{Statistical power (1 - probability of type II error)}

\item{tails}{One or two tailed test}
}
\value{
Minimum Detectable Effect Size
}
\description{
MDES for 2-Level Constant Effects Blocked Individual Random Assignment Designs
}

