% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_formulas.R
\name{BIRA2_1r}
\alias{BIRA2_1r}
\title{Model 2.3:  MDES Calculator for 2-Level Random Effects Blocked Individual Random Assignment (BIRA2_1r) Designs— Individuals Randomized within Blocks}
\usage{
BIRA2_1r(R2_1, R2_2T, P, g, n, J, rho, omega, alpha = 0.05, power = 0.8,
  tails = 2)
}
\arguments{
\item{R2_1}{Proportion of variance in level 1 outcome explained by level 1 covariates}

\item{R2_2T}{Proportion of between block variance in treatment effect explained by Level 2 covariates}

\item{P}{Proportion of level 1 units randomized to the treatment}

\item{g}{Number of level 2 covariates}

\item{n}{Average Block Size (mean # of level 1 units per level 2 unit)}

\item{J}{Number of level 2 units}

\item{rho}{proportion of variance in outcome between clusters}

\item{omega}{Treatment effect heterogeneity:  variability in treatment effects across Level 2 units, standardized by the variability in the Level-2 outcome}

\item{alpha}{Probability of a type I error}

\item{power}{Statistical power (1 - probability of type II error)}

\item{tails}{One or two tailed test}
}
\value{
Minimum Detectable Effect Size
}
\description{
MDES for 2-Level Random Effects Blocked Individual Random Assignment
}

