% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_formulas.R
\name{CRA2_2r}
\alias{CRA2_2r}
\title{Model 3.1:  MDES Calculator for Two-Level Cluster Random Assignment Design (CRA2_2)— Treatment at Level 2}
\usage{
CRA2_2r(R2_1, R2_2T, P, g, n, J, rho, alpha = 0.05, power = 0.8,
  tails = 2)
}
\arguments{
\item{R2_1}{Proportion of variance in level 1 outcome explained by level 1 covariates}

\item{R2_2T}{Proportion of variance in level 2 outcome explained by level 2 covariates}

\item{P}{Proportion of the level 2 units randomized to the treatment}

\item{g}{Number of level 2 Covariates used}

\item{n}{Mean # of level 1 units per level 2 cluster}

\item{J}{Number of level 2 units}

\item{rho}{proportion of variance in outcome that is between clusters}

\item{alpha}{Probability of a type I error}

\item{power}{Statistical power (1 - probability of type II error)}

\item{tails}{One or two tailed test}
}
\value{
Minimum Detectable Effect Size
}
\description{
MDEs for Two-Level Cluster Random Assignment Design
}

