% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_formulas.R
\name{CRA4_4r}
\alias{CRA4_4r}
\title{Model 3.3:  MDES Calculator for 4-Level Cluster Random Assignment Designs (CRA4_4r)—Treatment at Level 4}
\usage{
CRA4_4r(R2_1, R2_2, R2_3, R2_4, P, g_4, n, J, K, L, rho_2, rho_3, rho_4,
  alpha = 0.05, power = 0.8, tails = 2)
}
\arguments{
\item{R2_1}{Proportion of variance in level 1 outcome explained by level 1 covariates}

\item{R2_2}{Proportion of variance in level 2 outcome explained by level 2 covariates}

\item{R2_3}{Proportion of variance in level 3 outcome explained by level 3 covariates}

\item{R2_4}{Proportion of variance in level 4 outcome explained by level 4 covariates}

\item{P}{Proportion of the level 4 units randomized to the treatment}

\item{g_4}{Number of level 4 Covariates used}

\item{n}{Mean # of level 1 units per level 2 unit}

\item{J}{Mean # of level 2 units per level 3 unit}

\item{K}{Number of level 3 units per level 4 unit}

\item{L}{Number of level 4 units}

\item{rho_2}{proportion of variance in outcome that is between level 2 units}

\item{rho_3}{proportion of variance in outcome that is between level 3 units}

\item{rho_4}{proportion of variance in outcome that is between level 4 units}

\item{alpha}{Probability of a type I error}

\item{power}{Statistical power (1 - probability of type II error)}

\item{tails}{One or two tailed test}
}
\value{
Minimum Detectable Effect Size
}
\description{
MDES for 4-Level Cluster Random Assignment Designs
}

