% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis_functions.R
\name{get_alpha}
\alias{get_alpha}
\title{Calculate the level of significance.}
\usage{
get_alpha(power = 0.8, ICC = 0.001, delta = 0.2, M = 6, n = 100,
  type = "standard", cv = NULL, tol = .Machine$double.eps^0.25)
}
\arguments{
\item{power}{The power.}

\item{ICC}{The intra class correlation.}

\item{delta}{The effect size.}

\item{M}{The total number of clusters.}

\item{n}{The mean cluster size.}

\item{type}{The type of design effect, either standard or based on
the coefficient of variation.}

\item{cv}{The coefficient of variation, if type is cv.}

\item{tol}{A tolerance value.}
}
\value{
alpha The level of significance.
}
\description{
Calculate the level of significance.
}

