% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis_functions.R
\name{get_delta}
\alias{get_delta}
\title{Calculate the standardized effect size, delta.}
\usage{
get_delta(alpha = 0.05, power = 0.8, ICC = 0.001, M = 6, n = 100,
  type = "standard", cv = NULL, tol = .Machine$double.eps^0.25)
}
\arguments{
\item{alpha}{The level of significance.}

\item{power}{The power.}

\item{ICC}{The intra class correlation.}

\item{M}{The total number of clusters.}

\item{n}{The mean cluster size.}

\item{type}{The type of design effect, either standard or based on
the coefficient of variation.}

\item{cv}{The coefficient of variation, if type is cv.}

\item{tol}{A tolerance value.}
}
\value{
delta The effect size.
}
\description{
Calculate the standardized effect size, delta.
}

