% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis_functions.R
\name{get_M}
\alias{get_M}
\title{Calculate the total number of clusters.}
\usage{
get_M(alpha = 0.05, power = 0.8, ICC = 0.001, delta = 0.2, n = 100,
  type = "standard", cv = NULL, tol = .Machine$double.eps^0.25)
}
\arguments{
\item{alpha}{The level of significance.}

\item{power}{The power.}

\item{ICC}{The intra class correlation.}

\item{delta}{The effect size.}

\item{n}{The mean cluster size.}

\item{type}{The type of design effect, either standard or based on
the coefficient of variation.}

\item{cv}{The coefficient of variation, if type is cv.}

\item{tol}{A tolerance value.}
}
\value{
M The total number of clusters.
}
\description{
Calculate the total number of clusters.
}

