% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis_functions.R
\name{get_power}
\alias{get_power}
\title{Calculate the power of a cluster randomized design.}
\usage{
get_power(alpha = 0.05, delta, ICC, M, n, type = "standard", cv = NULL)
}
\arguments{
\item{alpha}{The level of significance.}

\item{delta}{The effect size.}

\item{ICC}{The intra-class correlation.}

\item{M}{The total number of clusters.}

\item{n}{The mean cluster size.}

\item{type}{The type of design effect, either standard or based on
the coefficient of variation.}

\item{cv}{The coefficient of variation, if type is cv.}
}
\value{
pwr The calculated power.
}
\description{
Calculate the power of a cluster randomized design.
}

