% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_formulas.R
\name{IRA}
\alias{IRA}
\title{Model 1.0: Individual Random assignment}
\usage{
IRA(R2, P, k, n, alpha = 0.05, power = 0.8, tails = 2)
}
\arguments{
\item{R2}{Proportion of variance explained in the outcome by the covariates}

\item{P}{Proportion of the sample randomized to the treatment}

\item{k}{Number of Covariates used}

\item{n}{Total Sample Size}

\item{alpha}{Probability of a type I error}

\item{power}{Statistical power (1 - probability of type II error)}

\item{tails}{One or two tailed test}
}
\value{
Minimum Detectable Effect Size
}
\description{
MDES for Individual Random assignment
}

