% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_power_sim.R
\name{crtPowerSim}
\alias{crtPowerSim}
\title{Calculate the p-value for a cluster randomized trial by simulation.}
\usage{
crtPowerSim(M = 10, n = 10, a = 10, vare = 0.95, varb = 0.05,
  B0 = 0, B1 = 0.2)
}
\arguments{
\item{M}{The total number of clusters.}

\item{n}{The mean cluster size.}

\item{a}{The minimum of the uniform distribution used to generate cluster sizes.}

\item{vare}{The variance of the measurement error.}

\item{varb}{The variance of the random effects.}

\item{B0}{The overall intercept.}

\item{B1}{The standardized treatment effect.}
}
\value{
The p-value under the null hypothesis.
}
\description{
Calculate the p-value for a cluster randomized trial by simulation.
}

