% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_crt_test.R
\name{power.crt.test}
\alias{power.crt.test}
\title{Power calculations for simple cluster randomized trials}
\usage{
power.crt.test(alpha = 0.05, power = 0.8, M = NULL, n = NULL, d = 0.2,
  ICC = NULL, type = "standard", cv = NULL,
  tol = .Machine$double.eps^0.25)
}
\arguments{
\item{alpha}{The level of significance of the test, the probability of a Type I error.}

\item{power}{The power of the test, 1 minus the probability of a Type II error.}

\item{M}{The total number of clusters.}

\item{n}{The mean cluster size.}

\item{d}{The standardized effect size.}

\item{ICC}{The intra-class correlation.}

\item{type}{The type of design effect, either 'standard' or based on
the coefficient of variation 'cv'.}

\item{cv}{The coefficient of variation. Used when 'type' is 'cv'.}

\item{tol}{Numerical tolerance used in root finding. The default providing
(at least) four significant digits.}
}
\value{
The computed argument.
}
\description{
Compute the power of a simple cluster randomized trial, or determine
  parameters to obtain a target power.
}

